+++ sampo/utilities/inseparables.py
@@ -1,5 +1,5 @@
-from sampo.schemas.graph import GraphNode
-from sampo.schemas.time import Time
+from sampo.schemas import GraphNode, ScheduledWork, Time, Worker, WorkTimeEstimator
+from sampo.schemas.schedule_spec import WorkSpec
 
 
 def get_exec_times_from_assigned_time_for_chain(inseparable_chain: list[GraphNode],
@@ -56,4 +56,54 @@ def get_exec_times_from_assigned_time_for_chain(inseparable_chain: list[GraphNod
     # The last node receives all remaining time to account for any rounding errors
     # during the distribution to previous nodes.
     exec_times[inseparable_chain[-1]] = remaining_time
+    return exec_times
+
+
+def find_min_time_slot_size(inseparable_chain: list[GraphNode],
+                            node2swork: dict[GraphNode, ScheduledWork],
+                            exec_times: dict[GraphNode, Time],
+                            start_time: Time) -> Time:
+    cur_finish_time = start_time
+    for dep_node in inseparable_chain:
+        # set start time as finish time of original work
+        # set finish time as finish time + working time of current node with identical resources
+        # (the same as in original work)
+        # set the same workers on it
+        # TODO Decide where this should be
+        dep_parent_time = dep_node.min_start_time(node2swork)
+
+        dep_st = max(cur_finish_time, dep_parent_time)
+
+        working_time = exec_times[dep_node]
+
+        cur_finish_time = dep_st + working_time
+
+    return cur_finish_time - start_time
+
+
+def calculate_exec_times(inseparable_chain: list[GraphNode],
+                         spec: WorkSpec,
+                         worker_team: list[Worker],
+                         work_estimator: WorkTimeEstimator) -> dict[GraphNode, Time]:
+    # TODO Refactor
+    spec_times = {}
+    if spec.assigned_time:
+        spec_times = get_exec_times_from_assigned_time_for_chain(inseparable_chain, spec.assigned_time)
+
+        assert sum(spec_times.values()) == spec.assigned_time
+
+    # 2. calculating execution time of the task
+
+    exec_times: dict[GraphNode, Time] = {}  # node: (lag, exec_time)
+    for chain_node in inseparable_chain:
+        # node_exec_time: Time = Time(0) if len(chain_node.work_unit.worker_reqs) == 0 else \
+        #     work_estimator.estimate_time(chain_node.work_unit, worker_team)
+        if spec.assigned_time:
+            node_exec_time = spec_times[chain_node]
+        else:
+            node_exec_time: Time = Time(0) if len(chain_node.work_unit.worker_reqs) == 0 else \
+                work_estimator.estimate_time(chain_node.work_unit, worker_team)
+
+        exec_times[chain_node] = node_exec_time
+
     return exec_times
